Lab09 - Forms
Added a form to sales.html with one text input and three number inputs.

Set up an addEventListener and handler function. Table re-renders each
time a new store is added.

If new store's name matches an already present one, that store is
updated with the new numbers and table is re-rendered.

Substantial refactoring of code overall.